docker_scripts/finetune_lighton_ocr/requirements.txt
torch==2.5.1+cu124
transformers>=5.0.0
datasets
evaluate
jiwer
pillow
pypdfium2
bitsandbytes
peft
accelerate
huggingface_hub
python-dotenv
supabase
tqdm
scikit-learn
sentencepiece
tiktoken
pillow-avif-plugin
# ── ONNX export ───────────────────────────────────────────────
optimum[exporters]
onnx
onnxruntime
